Supporting data and code upload
Data and R scripts relating to the methods described in McCarthy A, Caravaggi A,Fernández-Bellon D, Irwin S, Lusby J, O'Halloran J (*in review*) Bird and small mammal community composition and abundance in upland open habitats and early conifer forests.

These data and code have been archived and are accessible in order to support the manuscript in review and will be assigned a citable doi upon publication. They are not to be accessed or used prior to the confirmation of publication without the express permission of the authors.